requirements.txt
pyautogui==0.9.54
pytesseract==0.3.10
Pillow==10.0.0
PyGetWindow==0.0.9
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1